Django/Lift_Louisiana/requirements.txt
django
djangorestframework
docutils==0.14
jmespath==0.9.3
pyasn1==0.4.2
PyYAML==3.12
rsa==3.4.2
s3transfer==0.1.12
